Migrated chart legend to css modules
