--- src/components/ChartLegend/index.scss
@@ -1,117 +1,0 @@
-@use '../../styles' as *;
-
-.legend-container {
-  width: $chart-legend-width;
-  padding: $padding-component 0.25rem;
-  background-color: $color-component-bg;
-  height: $chart-legend-height;
-  @include default_border();
-
-  .legend-content {
-    height: 100%;
-    overflow-y: auto;
-
-    .legend-loading {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      height: 100%;
-
-      @media (max-width: $breakpoint-tablet) {
-        margin: 0.625rem 0;
-      }
-    }
-
-    ul {
-      margin: 0;
-      padding: 0;
-      list-style: none;
-    }
-
-    li {
-      padding: 0.5rem;
-      display: flex;
-      align-items: center;
-      transition: opacity 0.2s ease;
-      white-space: nowrap;
-      cursor: pointer;
-
-      &:hover {
-        background-color: #2a2a2a;
-      }
-
-      .color-marker {
-        display: inline-block;
-        width: 1rem;
-        height: 1rem;
-        margin-right: 0.5rem;
-        border-radius: 2px;
-        flex-shrink: 0;
-      }
-
-      .player-info {
-        color: #ffffff;
-        display: flex;
-        justify-content: space-between;
-        width: 100%;
-        min-width: 0;
-
-        .player-name-container {
-          display: flex;
-          min-width: 0;
-          margin-right: 0.375rem;
-
-          .player-name {
-            overflow: hidden;
-            text-overflow: ellipsis;
-            white-space: nowrap;
-
-            &.first-place {
-              font-weight: bold;
-            }
-          }
-
-          .special-icon {
-            flex-shrink: 0;
-            margin-left: 0.25rem;
-            line-height: 1;
-            font-size: 1.2em;
-            position: relative;
-            @include rotateY360Animation('auto');
-          }
-
-          &.has-special-icon:hover .special-icon {
-            transform: rotateY(360deg);
-          }
-        }
-
-        .player-time {
-          margin-left: 0.375rem;
-          opacity: 0.7;
-          flex-shrink: 0;
-
-          &.first-place {
-            font-weight: bold;
-            opacity: 1;
-          }
-        }
-      }
-    }
-  }
-
-  @media (max-width: $breakpoint-tablet) {
-    width: 100%;
-    height: auto;
-
-    ul {
-      display: flex;
-      flex-wrap: wrap;
-      gap: 0.5rem;
-    }
-
-    li {
-      flex: 1;
-      min-width: 12.5rem;
-    }
-  }
-}
